Car: Entity, dto, CRUDService, IService, Rest controller, controller Test, Repository Driver: Entity, dto, CRUDService, IService, Rest controller Test, Repository

+++ src/main/java/com/diploma/maksimov/db/repository/CarRepository.java
@@ -3,5 +3,5 @@ package com.diploma.maksimov.db.repository;
 import com.diploma.maksimov.db.entity.CarEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarRepository extends CrudRepository<CarEntity,Integer> {
+public interface CarRepository extends CrudRepository<CarEntity,Long> {
 }
